+++ visidata/mainloop.py
@@ -27,6 +27,7 @@ def draw_sheet(self, scr, sheet):
     sheet.ensureLoaded()
 
     scr.erase()  # clear screen before every re-draw
+    scr.bkgd(' ', colors.color_default)
 
     sheet._scr = scr
 
@@ -85,7 +86,6 @@ def draw_all(vd):
     ss2 = vd.sheetstack(2)
     if ss1 and not ss2:
         vd.activePane = 1
-        ss1[0].options.disp_splitwin_pct = 0
         vd.setWindows(vd.scrFull)
         vd.draw_sheet(vd.win1, ss1[0])
         if vd.win2:
@@ -93,12 +93,11 @@ def draw_all(vd):
             vd.win2.refresh()
     elif not ss1 and ss2:
         vd.activePane = 2
-        ss2[0].options.disp_splitwin_pct = 0
         vd.setWindows(vd.scrFull)
-        vd.draw_sheet(vd.win1, ss2[0])
-        if vd.win2:
-            vd.win2.erase()
-            vd.win2.refresh()
+        vd.draw_sheet(vd.win2, ss2[0])
+        if vd.win1:
+            vd.win1.erase()
+            vd.win1.refresh()
     elif ss1 and ss2 and vd.win2:
         vd.draw_sheet(vd.win1, ss1[0])
         vd.draw_sheet(vd.win2, ss2[0])